refactor: extract execution guards module
