+++ projects/04-llm-adapter/adapter/core/runner_execution_attempts.py
@@ -16,6 +16,13 @@ if TYPE_CHECKING:  # pragma: no cover - 型補完用
     from .runner_api import RunnerConfig
     from .runner_execution import SingleRunResult
 
+    _RunSingle = Callable[
+        [ProviderConfig, BaseProvider, GoldenTask, int, str],
+        SingleRunResult,
+    ]
+else:
+    _RunSingle = Callable[[ProviderConfig, BaseProvider, GoldenTask, int, str], object]
+
 
 _single_run_result_cls: type[Any] | None = None
 
@@ -37,12 +44,6 @@ class _ParallelRunner(Protocol):
         max_concurrency: int | None = None,
     ) -> object:
         ...
-
-
-_RunSingle = Callable[
-    [ProviderConfig, BaseProvider, GoldenTask, int, str],
-    SingleRunResult,
-]
 
 
 class SequentialAttemptExecutor:
